refactored to move knowledge of BanyPlugin to outermost layer
- introduced new interface that is used by instances of
  ThirdPartyTransactionGateway. Implementations wrap BanyPlugin
  instances
- refactored to use OffsetDateTime instead of LocalDateTime
- refactored CibcTransaction to use BigDecimal under the hood

+++ plugins/cibc/src/main/kotlin/com/pissiphany/bany/plugin/cibc/CibcTransactionService.kt
@@ -116,6 +116,8 @@ internal class CibcTransactionService(
     }
 
     override fun tearDown() {
+        if (token.isBlank()) return
+
         val request = Request.Builder()
             .url(env.authUrl)
             .delete()
